chorde(deep_causality_haft); reorganized source code
