+++ deep_causality_haft/src/morphism/morphism_endo.rs
